Hardening/1368 cppcheck01

+++ src/lib/serviceRoutinesV2/postSubscriptions.cpp
@@ -57,8 +57,8 @@ extern std::string postSubscriptions
     const size_t  MSG_SIZE        = 96;  // strlen(msg) + enough room for digits
     char          errMsg[MSG_SIZE];
 
-    snprintf(errMsg, MSG_SIZE, "max *one* service-path allowed for subscriptions (%zd given)",
-             ciP->servicePathV.size());
+    snprintf(errMsg, MSG_SIZE, "max *one* service-path allowed for subscriptions (%lu given)",
+             (unsigned long) ciP->servicePathV.size());
     alarmMgr.badInput(clientIp, errMsg);
     scr.subscribeError.errorCode.fill(SccBadRequest, "max one service-path allowed for subscriptions");
